requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyJWT==2.9.0
pytest==8.3.2
httpx==0.27.2